✨ mvc 구조로 파일 분리
