+++ src/routes/chat.route.js
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { handleChat } from "../controllers/chat.controller.js";
+
+const chatRouter = Router();
+
+chatRouter.post("/", handleChat);
+
+export default chatRouter;
