+++ src/routes/index.js
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import chatRouter from "./chat.route.js";
+
+const router = Router();
+
+router.use("/chat", chatRouter);
+
+export default router;
